Navigation in home is done, created two packages: popular and services for comfortable
