+++ app/src/main/java/com/example/yourlicey28/presentation/home/components/popular/my_teachers.kt
@@ -1,0 +1,10 @@
+package com.example.yourlicey28.presentation.home.components.popular
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun MyTeachersScreen(navController: NavHostController){
+    Text(text = "Мои учителя")
+}
