+++ app/src/main/java/com/example/yourlicey28/presentation/home/components/popular/time_table.kt
@@ -1,0 +1,10 @@
+package com.example.yourlicey28.presentation.home.components.popular
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun TimeTableScreen(navController: NavHostController){
+    Text(text = "Раписание")
+}
