+++ app/src/main/java/com/example/yourlicey28/presentation/home/components/services/enter_to_first_class.kt
@@ -1,0 +1,10 @@
+package com.example.yourlicey28.presentation.home.components.services
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun EnterToFirstClass(navController: NavHostController){
+    Text(text = "Приём в 1 класс")
+}
