+++ app/src/main/java/com/example/yourlicey28/presentation/home/components/services/photogallery.kt
@@ -1,0 +1,10 @@
+package com.example.yourlicey28.presentation.home.components.services
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun PhotoGallery(navController: NavHostController){
+    Text(text = "Фотогаллерея")
+}
